# feat: count underscore in English tokenizer and improve performance