I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)

## What's changed and what's your intention?

- don't separate on '_'
- performance improvement
- benchmark

Benchmark compare with previous implementation:
```
english_tokenizer/tokenize/short
                        time:   [36.693 ns 36.861 ns 37.038 ns]
                        change: [-32.648% -32.371% -32.079%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
english_tokenizer/tokenize/medium
                        time:   [153.71 ns 154.05 ns 154.40 ns]
                        change: [-38.949% -38.758% -38.523%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
english_tokenizer/tokenize/long
                        time:   [416.41 ns 418.59 ns 420.88 ns]
                        change: [-54.355% -54.151% -53.955%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) low mild
english_tokenizer/tokenize/mixed_chars
                        time:   [96.234 ns 96.810 ns 97.442 ns]
                        change: [-53.532% -53.339% -53.141%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
english_tokenizer/tokenize/numbers_heavy
                        time:   [101.51 ns 101.93 ns 102.41 ns]
                        change: [-34.777% -34.591% -34.387%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) high mild
  11 (11.00%) high severe
english_tokenizer/tokenize/punctuation_heavy
                        time:   [102.37 ns 102.73 ns 103.12 ns]
                        change: [-44.286% -44.103% -43.901%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
english_tokenizer/tokenize/postgres log
                        time:   [190.18 ns 190.94 ns 191.73 ns]
                        change: [-53.639% -53.528% -53.421%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  2 (2.00%) low mild
  5 (5.00%) high mild
  4 (4.00%) high severe
english_tokenizer/tokenize/many_short_words
                        time:   [126.79 ns 127.11 ns 127.38 ns]
                        change: [-23.708% -23.345% -22.893%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
english_tokenizer/tokenize/with_unicode
                        time:   [61.698 ns 62.185 ns 62.698 ns]
                        change: [-73.410% -73.244% -73.082%] (p = 0.00 < 0.05)
                        Performance has improved.

english_tokenizer_repeated/repeated_tokenize/10
                        time:   [1.1401 µs 1.1422 µs 1.1445 µs]
                        change: [-37.271% -37.083% -36.902%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
english_tokenizer_repeated/repeated_tokenize/100
                        time:   [11.318 µs 11.349 µs 11.383 µs]
                        change: [-35.162% -34.549% -33.968%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
english_tokenizer_repeated/repeated_tokenize/1000
                        time:   [99.466 µs 99.905 µs 100.36 µs]
                        change: [-37.585% -37.377% -37.148%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild
```

## PR Checklist
Please convert it to a draft if some of the following conditions are not met.

- [x] I have written the necessary rustdoc comments.
- [x] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.
- [ ] API changes are backward compatible.
- [ ] Schema or data changes are backward compatible.