@zhongzc 2025-08-06
https://github.com/GreptimeTeam/greptimedb/blob/414db41219916c99cc3e75bf56121424fbd3c85d/src/query/src/optimizer/constant_term.rs#L180-L185

Need to adjust here